% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{EM Algorithm for Fitting the Hidden Markov Model}
\usage{
EM(initial.value, obs)
}
\arguments{
\item{initial.value}{Output from initial.value function.}

\item{obs}{Output from read.process function.}
}
\value{
res A matrix contains tansformed observations for each bin and the predicted direction for each bin (0-normal, 1-hyper, 2-hypo).

para Fitted parameters for HMM.
}
\description{
This function implements the EM algorithm to estimate HMM parameters,
and find the best sequence of hidden states based on model fitting.
}
