% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialValue.R
\name{initial.value}
\alias{initial.value}
\title{Set up Initial Value for EM algorithm}
\usage{
initial.value(obs)
}
\arguments{
\item{obs}{Output from read.process function}
}
\value{
initial.value Initial values for EM algorithm
}
\description{
This function sets up initial values for estimating hidden Markov model 
parameters by EM algorithm
}
