% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.process.R
\name{read.process}
\alias{read.process}
\title{Read and Transform Observed Sequencing Data}
\usage{
read.process(pos, norm.m, norm.um, abnorm.m, abnorm.um, bin.size = 40)
}
\arguments{
\item{pos}{CpG positions.}

\item{norm.m}{A matrix contains methylated read count data of normal group. Each column of the matrix represents a replicate and each row represents a CpG position.}

\item{norm.um}{A matrix contains unmethylated read count data of normal group.}

\item{abnorm.m}{A matrix contains methylated read count data of abnormal group.}

\item{abnorm.um}{A matrix contains unmethylated read count data of abnormal group.}

\item{bin.size}{Size of a bin. Default to 40.}
}
\value{
obs A matrix contains transformed observations for each bin, distance between the current bin the bin ahead of it, average methylation rate of abnormal and normal groups, start and end position of each bin.
}
\description{
This function reads in methylated and unmethylated read count data,
and transform it into logarithm bin-wise data.
}
