% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM3.R
\name{EM}
\alias{EM}
\title{EM Algorithm for HMM}
\usage{
EM(initial.value, obs)
}
\arguments{
\item{initial.value}{Initial values for EM algorithm. This is the output from initial.value function}

\item{obs}{Transformed observations for all bins. This is the output from read.process function}
}
\description{
This function implements the EM algorithm for estimating HMM parameters.
}
