% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialValue2.R
\name{initial.value}
\alias{initial.value}
\title{Set up Initial Values for HMM}
\usage{
initial.value(obs)
}
\arguments{
\item{obs}{Transformed observations for bins. This is output from read.process function}
}
\description{
This function sets up initial values for the EM algorithm to
estimate parameters of the hidden Markov model.
}
