% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostAdjustment.R
\name{PostAdjustment}
\alias{PostAdjustment}
\title{Post Adjustment}
\usage{
PostAdjustment(obs, em.o, CpG.pos, min.length = 200, min.CpGs = 3)
}
\arguments{
\item{obs}{Transformed observations for all bins. This is the output from read.process function}

\item{em.o}{Parameter estimation and output from HMM}

\item{CpG.pos}{CpG positions}

\item{min.length}{User-defined minimum DMR length. Default to 200}

\item{min.CpGs}{User-defined minimum number of CpGs in each DMR. Default to 3}
}
\description{
This function selects output from HMM to ensure the minimum length of DMRs and 
minimum number of CpGs in each DMR.
}
