% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostAdjustment.R
\name{PostAdjustment}
\alias{PostAdjustment}
\title{Post Adjustment}
\usage{
PostAdjustment(em.o, CpG.pos, min.length = 200, min.CpGs = 3)
}
\arguments{
\item{em.o}{Output from EM function.}

\item{CpG.pos}{Contains all CpG positions.}

\item{min.length}{Minimum length of a DMR. Default to 200.}

\item{min.CpGs}{Minimum number of CpGs contained in a DMR. Default to 3.}
}
\value{
dmr.res A matrix contains transformed observations for each bin and the adjusted predictions for each bin (0-normal, 1-hyper, 2-hypo).

region A matrix contains detected regions satisfying user-defined length and number of CpGs.
}
\description{
This function adjusts HMM output such that each detected DMR has a minimum length and a minimum number of CpGs in each DMR.
}
